requirements.txt
langchain
langchain-community
langchain-core 
google-generativeai
langchain_google_genai 
python-dotenv
streamlit
sentence-transformers
beautifulsoup4
bs4
faiss-cpu
pypdf